refactor transfer to mysql

+++ src/main/resources/db/migration/V5__create_comment_table.sql
@@ -3,7 +3,7 @@ create table comment
     id           bigint auto_increment,
     parent_id    bigint           not null,
     type         int              not null,
-    commentator  int              not null,
+    commentator  BIGINT           not null,
     gmt_creat    bigint           not null,
     gmt_modified bigint           not null,
     like_count   bigint default 0 not null,
